Update scripts.
* When configuring sudoers we try to add to the include dir if that's defined on the configuration file, if not, just append the data.
* When cleaning up, remove virtualbox stuff if `.vbox_version` is present.

+++ definitions/.common/remove-temporary-files.sh
@@ -14,9 +14,11 @@ set -o nounset                              # Treat unset variables as an error
 echo
 echo "Cleaning up temporary files"
 
-VBOX_VERSION=$(cat .vbox_version)
-VBOX_ISO=VBoxGuestAdditions_$VBOX_VERSION.iso
-test -f $VBOX_ISO && rm -f $VBOX_ISO
+if [ -f .vbox_version ]; then
+    VBOX_VERSION=$(cat .vbox_version)
+    VBOX_ISO=VBoxGuestAdditions_$VBOX_VERSION.iso
+    test -f $VBOX_ISO && rm -f $VBOX_ISO
+fi
 rm -f /tmp/*.sh
 rm -f /home/vagrant/*.sh
 rm -f /home/vagrant/.vbox_version
